requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic-settings==2.6.0
